pyproject.toml
[tool.poetry]
name = "lastfm"
version = "0.1.0"
description = ""
authors = ["tadamcz <tadamczewskipublic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pylast = "^5.1.0"
python-dotenv = "^1.0.0"
humanize = "^4.6.0"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^23.1.0"}

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

